Move rumEventsCollection/error
