+++ packages/rum-core/src/boot/startRum.ts
@@ -17,7 +17,7 @@ import { LifeCycle, LifeCycleEventType } from '../domain/lifeCycle'
 import { startViewContexts } from '../domain/contexts/viewContexts'
 import { startRequestCollection } from '../domain/requestCollection'
 import { startActionCollection } from '../domain/action/actionCollection'
-import { startErrorCollection } from '../domain/rumEventsCollection/error/errorCollection'
+import { startErrorCollection } from '../domain/error/errorCollection'
 import { startLongTaskCollection } from '../domain/rumEventsCollection/longTask/longTaskCollection'
 import { startResourceCollection } from '../domain/rumEventsCollection/resource/resourceCollection'
 import { startViewCollection } from '../domain/rumEventsCollection/view/viewCollection'
