+++ packages/rum-core/src/domain/error/errorCollection.spec.ts
@@ -1,10 +1,10 @@
 import type { RelativeTime, TimeStamp, ErrorWithCause } from '@datadog/browser-core'
 import { ErrorHandling, ErrorSource, NO_ERROR_STACK_PRESENT_MESSAGE } from '@datadog/browser-core'
-import type { TestSetupBuilder } from '../../../../test'
-import { setup } from '../../../../test'
-import type { RawRumErrorEvent } from '../../../rawRumEvent.types'
-import { RumEventType } from '../../../rawRumEvent.types'
-import { LifeCycleEventType } from '../../lifeCycle'
+import type { TestSetupBuilder } from '../../../test'
+import { setup } from '../../../test'
+import type { RawRumErrorEvent } from '../../rawRumEvent.types'
+import { RumEventType } from '../../rawRumEvent.types'
+import { LifeCycleEventType } from '../lifeCycle'
 import { doStartErrorCollection } from './errorCollection'
 
 describe('error collection', () => {
