+++ packages/rum-core/src/domain/error/trackConsoleError.spec.ts
