+++ packages/rum-core/src/domain/error/trackConsoleError.ts
