+++ packages/rum-core/src/domain/error/trackReportError.ts
@@ -1,6 +1,6 @@
 import type { Observable, RawError } from '@datadog/browser-core'
 import { clocksNow, ErrorHandling, ErrorSource, initReportObservable, RawReportType } from '@datadog/browser-core'
-import type { RumConfiguration } from '../../configuration'
+import type { RumConfiguration } from '../configuration'
 
 export function trackReportError(configuration: RumConfiguration, errorObservable: Observable<RawError>) {
   const subscription = initReportObservable(configuration, [
